requirements.txt
pandas>=1.3.0
python-dateutil>=2.8.2
pytz>=2021.3